Think about alternative way to achieve this
I'm confused by the way this website is set up because if I'm offering help, I help someone can see my offer, but the way that this website is set up, people who are seeking will create a post that is seeking for help. Somehow this is defeating the purpose of the offer/seek, because these two users will not cross each other. People will post on seek or post on offer. 
Overall, I love the idea, it is something that I wanted to do as a side project before.